fix: add exception and adjust to R2
'Immunization.occurrenceDateTime" did not exist in R2. Since the Test
server is still on R2, only 'Immunization.expirationDate' is supported
for extracting a relevant date time. Also, a new exception is
introduced, so when there is no 'Immunization.expirationDate', but
'Immunization.status' is set to 'completed', the exception is thrown.
The service can then catch it and return an appropriate response.